added hash from config to lockfile
this way you can run 2 completely different config in parallel
whilst avoiding running identical configs in parallel.
This will partially circumvent problems reading and writing to the sqlite
but it's not 100% foolproof. One can still use 2 configs using the same db
whilst having any other parameter different and run them in parallel.
In this case, i consider this to be a PICNIC problem